+++ packages/core/cache/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @atlaspack/cache
 
+## 3.2.18
+
+### Patch Changes
+
+- Updated dependencies [[`23d561e`](https://github.com/atlassian-labs/atlaspack/commit/23d561e51e68b0c38fd1ff4e4fb173e5e7b01cf2)]:
+  - @atlaspack/feature-flags@2.21.0
+  - @atlaspack/utils@2.18.0
+  - @atlaspack/rust@3.5.0
+  - @atlaspack/fs@2.15.18
+  - @atlaspack/logger@2.14.15
+
 ## 3.2.17
 
 ### Patch Changes