+++ packages/optimizers/inline-requires/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atlaspack/optimizer-inline-requires
 
+## 2.14.23
+
+### Patch Changes
+
+- Updated dependencies [[`23d561e`](https://github.com/atlassian-labs/atlaspack/commit/23d561e51e68b0c38fd1ff4e4fb173e5e7b01cf2)]:
+  - @atlaspack/feature-flags@2.21.0
+  - @atlaspack/rust@3.5.0
+  - @atlaspack/types@2.15.13
+  - @atlaspack/plugin@2.14.23
+
 ## 2.14.22
 
 ### Patch Changes