+++ packages/reporters/lsp-reporter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaspack/reporter-lsp
 
+## 2.14.23
+
+### Patch Changes
+
+- Updated dependencies [[`23d561e`](https://github.com/atlassian-labs/atlaspack/commit/23d561e51e68b0c38fd1ff4e4fb173e5e7b01cf2)]:
+  - @atlaspack/utils@2.18.0
+  - @atlaspack/plugin@2.14.23
+
 ## 2.14.22
 
 ### Patch Changes