+++ packages/runtimes/js/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @atlaspack/runtime-js
 
+## 2.17.0
+
+### Minor Changes
+
+- [#731](https://github.com/atlassian-labs/atlaspack/pull/731) [`23d561e`](https://github.com/atlassian-labs/atlaspack/commit/23d561e51e68b0c38fd1ff4e4fb173e5e7b01cf2) Thanks [@marcins](https://github.com/marcins)! - Implement "inline isolated" scripts
+
+### Patch Changes
+
+- Updated dependencies [[`23d561e`](https://github.com/atlassian-labs/atlaspack/commit/23d561e51e68b0c38fd1ff4e4fb173e5e7b01cf2)]:
+  - @atlaspack/feature-flags@2.21.0
+  - @atlaspack/utils@2.18.0
+  - @atlaspack/plugin@2.14.23
+
 ## 2.16.1
 
 ### Patch Changes